Fixes #577
Latest code review changes
