+++ extensions/servlet/src/main/java/com/stormpath/sdk/servlet/config/Config.java
@@ -18,7 +18,7 @@ package com.stormpath.sdk.servlet.config;
 import com.stormpath.sdk.authc.AuthenticationResult;
 import com.stormpath.sdk.servlet.filter.ControllerConfigResolver;
 import com.stormpath.sdk.servlet.http.Saver;
-import com.stormpath.sdk.servlet.mvc.Controller;
+import com.stormpath.sdk.servlet.http.authc.AccountStoreResolver;
 
 import javax.servlet.ServletException;
 import java.util.List;
@@ -43,7 +43,9 @@ public interface Config extends Map<String, String> {
 
     ControllerConfigResolver getChangePasswordControllerConfig();
 
-    Saver<AuthenticationResult> getAuthenticationResultSaver() throws ServletException;
+    Saver<AuthenticationResult> getAuthenticationResultSaver();
+
+    AccountStoreResolver getAccountStoreResolver();
 
     /**
      * @since 1.0.RC6
