+++ extensions/servlet/src/main/java/com/stormpath/sdk/servlet/mvc/RegisterController.java
@@ -26,9 +26,9 @@ import com.stormpath.sdk.resource.ResourceException;
 import com.stormpath.sdk.servlet.account.event.RegisteredAccountRequestEvent;
 import com.stormpath.sdk.servlet.account.event.impl.DefaultRegisteredAccountRequestEvent;
 import com.stormpath.sdk.servlet.authc.impl.TransientAuthenticationResult;
+import com.stormpath.sdk.servlet.config.Config;
 import com.stormpath.sdk.servlet.event.RequestEvent;
 import com.stormpath.sdk.servlet.event.impl.Publisher;
-import com.stormpath.sdk.servlet.filter.ControllerConfigResolver;
 import com.stormpath.sdk.servlet.form.Field;
 import com.stormpath.sdk.servlet.form.Form;
 import com.stormpath.sdk.servlet.http.Saver;
@@ -61,21 +61,15 @@ public class RegisterController extends FormController {
         super();
     }
 
-    public RegisterController(ControllerConfigResolver controllerConfigResolver,
-                              Client client,
-                              Publisher<RequestEvent> eventPublisher,
-                              List<Field> formFields,
-                              Saver<AuthenticationResult> authenticationResultSaver,
-                              String loginUri,
-                              String verifyViewName) {
-        super(controllerConfigResolver);
+    public RegisterController(Config config, Client client, Publisher<RequestEvent> eventPublisher, List<Field> formFields) {
+        super(config.getRegisterControllerConfig());
 
         this.client = client;
         this.eventPublisher = eventPublisher;
         this.formFields = formFields;
-        this.authenticationResultSaver = authenticationResultSaver;
-        this.loginUri = loginUri;
-        this.verifyViewName = verifyViewName;
+        this.authenticationResultSaver = config.getAuthenticationResultSaver();
+        this.loginUri = config.getLoginControllerConfig().getUri();
+        this.verifyViewName = config.getVerifyControllerConfig().getView();
 
         Assert.notNull(client, "client cannot be null.");
         Assert.notNull(eventPublisher, "eventPublisher cannot be null.");
