+++ extensions/spring/stormpath-spring-webmvc/src/main/java/com/stormpath/spring/config/AbstractStormpathWebMvcConfiguration.java
@@ -27,6 +27,7 @@ import com.stormpath.sdk.lang.Collections;
 import com.stormpath.sdk.lang.Strings;
 import com.stormpath.sdk.saml.SamlResultListener;
 import com.stormpath.sdk.servlet.authz.RequestAuthorizer;
+import com.stormpath.sdk.servlet.config.Config;
 import com.stormpath.sdk.servlet.config.CookieConfig;
 import com.stormpath.sdk.servlet.csrf.CsrfTokenManager;
 import com.stormpath.sdk.servlet.csrf.DefaultCsrfTokenManager;
@@ -37,20 +38,80 @@ import com.stormpath.sdk.servlet.event.RequestEventListenerAdapter;
 import com.stormpath.sdk.servlet.event.TokenRevocationRequestEventListener;
 import com.stormpath.sdk.servlet.event.impl.Publisher;
 import com.stormpath.sdk.servlet.event.impl.RequestEventPublisher;
-import com.stormpath.sdk.servlet.filter.*;
-import com.stormpath.sdk.servlet.filter.account.*;
-import com.stormpath.sdk.servlet.filter.oauth.*;
+import com.stormpath.sdk.servlet.filter.ControllerConfigResolver;
+import com.stormpath.sdk.servlet.filter.DefaultServerUriResolver;
+import com.stormpath.sdk.servlet.filter.DefaultUsernamePasswordRequestFactory;
+import com.stormpath.sdk.servlet.filter.DefaultWrappedServletRequestFactory;
+import com.stormpath.sdk.servlet.filter.FilterChainResolver;
+import com.stormpath.sdk.servlet.filter.ProxiedFilterChain;
+import com.stormpath.sdk.servlet.filter.ServerUriResolver;
+import com.stormpath.sdk.servlet.filter.StormpathFilter;
+import com.stormpath.sdk.servlet.filter.UsernamePasswordRequestFactory;
+import com.stormpath.sdk.servlet.filter.WrappedServletRequestFactory;
+import com.stormpath.sdk.servlet.filter.account.AccountResolverFilter;
+import com.stormpath.sdk.servlet.filter.account.AuthenticationJwtFactory;
+import com.stormpath.sdk.servlet.filter.account.AuthenticationResultSaver;
+import com.stormpath.sdk.servlet.filter.account.AuthorizationHeaderAccountResolver;
+import com.stormpath.sdk.servlet.filter.account.CookieAccountResolver;
+import com.stormpath.sdk.servlet.filter.account.CookieAuthenticationResultSaver;
+import com.stormpath.sdk.servlet.filter.account.DefaultAuthenticationJwtFactory;
+import com.stormpath.sdk.servlet.filter.account.DefaultJwtAccountResolver;
+import com.stormpath.sdk.servlet.filter.account.DefaultJwtSigningKeyResolver;
+import com.stormpath.sdk.servlet.filter.account.JwtAccountResolver;
+import com.stormpath.sdk.servlet.filter.account.JwtSigningKeyResolver;
+import com.stormpath.sdk.servlet.filter.account.SessionAccountResolver;
+import com.stormpath.sdk.servlet.filter.account.SessionAuthenticationResultSaver;
+import com.stormpath.sdk.servlet.filter.oauth.AccessTokenAuthenticationRequestFactory;
+import com.stormpath.sdk.servlet.filter.oauth.AccessTokenResultFactory;
+import com.stormpath.sdk.servlet.filter.oauth.DefaultAccessTokenAuthenticationRequestFactory;
+import com.stormpath.sdk.servlet.filter.oauth.DefaultAccessTokenRequestAuthorizer;
+import com.stormpath.sdk.servlet.filter.oauth.DefaultAccessTokenResultFactory;
+import com.stormpath.sdk.servlet.filter.oauth.DefaultRefreshTokenAuthenticationRequestFactory;
+import com.stormpath.sdk.servlet.filter.oauth.DefaultRefreshTokenResultFactory;
+import com.stormpath.sdk.servlet.filter.oauth.OriginAccessTokenRequestAuthorizer;
+import com.stormpath.sdk.servlet.filter.oauth.RefreshTokenAuthenticationRequestFactory;
+import com.stormpath.sdk.servlet.filter.oauth.RefreshTokenResultFactory;
 import com.stormpath.sdk.servlet.form.DefaultField;
 import com.stormpath.sdk.servlet.form.Field;
 import com.stormpath.sdk.servlet.http.InvalidMediaTypeException;
 import com.stormpath.sdk.servlet.http.MediaType;
 import com.stormpath.sdk.servlet.http.Resolver;
 import com.stormpath.sdk.servlet.http.Saver;
-import com.stormpath.sdk.servlet.http.authc.*;
+import com.stormpath.sdk.servlet.http.authc.AccountStoreResolver;
+import com.stormpath.sdk.servlet.http.authc.AuthorizationHeaderAuthenticator;
+import com.stormpath.sdk.servlet.http.authc.BasicAuthenticationScheme;
+import com.stormpath.sdk.servlet.http.authc.BearerAuthenticationScheme;
+import com.stormpath.sdk.servlet.http.authc.DisabledAccountStoreResolver;
+import com.stormpath.sdk.servlet.http.authc.HeaderAuthenticator;
+import com.stormpath.sdk.servlet.http.authc.HttpAuthenticationScheme;
+import com.stormpath.sdk.servlet.i18n.MessageTag;
 import com.stormpath.sdk.servlet.idsite.DefaultIdSiteOrganizationResolver;
 import com.stormpath.sdk.servlet.idsite.IdSiteOrganizationContext;
-import com.stormpath.sdk.servlet.mvc.*;
-import com.stormpath.sdk.servlet.mvc.provider.*;
+import com.stormpath.sdk.servlet.mvc.AccessTokenController;
+import com.stormpath.sdk.servlet.mvc.ChangePasswordController;
+import com.stormpath.sdk.servlet.mvc.DefaultFormFieldsParser;
+import com.stormpath.sdk.servlet.mvc.ErrorModelFactory;
+import com.stormpath.sdk.servlet.mvc.ForgotPasswordController;
+import com.stormpath.sdk.servlet.mvc.FormFieldParser;
+import com.stormpath.sdk.servlet.mvc.IdSiteController;
+import com.stormpath.sdk.servlet.mvc.IdSiteLogoutController;
+import com.stormpath.sdk.servlet.mvc.IdSiteResultController;
+import com.stormpath.sdk.servlet.mvc.LoginController;
+import com.stormpath.sdk.servlet.mvc.LoginErrorModelFactory;
+import com.stormpath.sdk.servlet.mvc.LogoutController;
+import com.stormpath.sdk.servlet.mvc.MeController;
+import com.stormpath.sdk.servlet.mvc.RegisterController;
+import com.stormpath.sdk.servlet.mvc.SamlController;
+import com.stormpath.sdk.servlet.mvc.SamlLogoutController;
+import com.stormpath.sdk.servlet.mvc.SamlResultController;
+import com.stormpath.sdk.servlet.mvc.SendVerificationEmailController;
+import com.stormpath.sdk.servlet.mvc.VerifyController;
+import com.stormpath.sdk.servlet.mvc.provider.AccountStoreModelFactory;
+import com.stormpath.sdk.servlet.mvc.provider.DefaultAccountStoreModelFactory;
+import com.stormpath.sdk.servlet.mvc.provider.FacebookCallbackController;
+import com.stormpath.sdk.servlet.mvc.provider.GithubCallbackController;
+import com.stormpath.sdk.servlet.mvc.provider.GoogleCallbackController;
+import com.stormpath.sdk.servlet.mvc.provider.LinkedinCallbackController;
 import com.stormpath.sdk.servlet.oauth.AccessTokenValidationStrategy;
 import com.stormpath.sdk.servlet.organization.DefaultOrganizationNameKeyResolver;
 import com.stormpath.sdk.servlet.saml.DefaultSamlOrganizationResolver;
@@ -60,7 +121,16 @@ import com.stormpath.sdk.servlet.util.RemoteAddrResolver;
 import com.stormpath.sdk.servlet.util.SecureRequiredExceptForLocalhostResolver;
 import com.stormpath.sdk.servlet.util.SubdomainResolver;
 import com.stormpath.spring.context.CompositeMessageSource;
-import com.stormpath.spring.mvc.*;
+import com.stormpath.spring.mvc.ChangePasswordControllerConfigResolver;
+import com.stormpath.spring.mvc.ForgotPasswordControllerConfigResolver;
+import com.stormpath.spring.mvc.LoginControllerConfigResolver;
+import com.stormpath.spring.mvc.LogoutControllerConfigResolver;
+import com.stormpath.spring.mvc.RegisterControllerConfigResolver;
+import com.stormpath.spring.mvc.SendVerificationEmailControllerConfigResolver;
+import com.stormpath.spring.mvc.SpringController;
+import com.stormpath.spring.mvc.SpringSpaController;
+import com.stormpath.spring.mvc.TemplateLayoutInterceptor;
+import com.stormpath.spring.mvc.VerifyControllerConfigResolver;
 import io.jsonwebtoken.SignatureAlgorithm;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -74,7 +144,9 @@ import org.springframework.context.support.ResourceBundleMessageSource;
 import org.springframework.core.Ordered;
 import org.springframework.util.PathMatcher;
 import org.springframework.util.StringUtils;
-import org.springframework.web.servlet.*;
+import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.HandlerMapping;
+import org.springframework.web.servlet.LocaleResolver;
 import org.springframework.web.servlet.View;
 import org.springframework.web.servlet.ViewResolver;
 import org.springframework.web.servlet.handler.SimpleUrlHandlerMapping;
@@ -90,7 +162,16 @@ import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.LinkedHashMap;
+import java.util.LinkedHashSet;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.Set;
 
 /**
  * @since 1.0.RC4
@@ -835,16 +916,7 @@ public abstract class AbstractStormpathWebMvcConfiguration {
         }
 
         //otherwise standard login controller:
-        LoginController controller = new LoginController(
-                stormpathLoginControllerConfigResolver(),
-                stormpathForgotPasswordControllerConfigResolver().getUri(),
-                stormpathVerifyControllerConfigResolver().getUri(),
-                stormpathRegisterControllerConfigResolver().getUri(),
-                stormpathLogoutControllerConfigResolver().getUri(),
-                stormpathVerifyControllerConfigResolver().isEnabled(),
-                stormpathAuthenticationResultSaver(),
-                loginErrorModelFactory
-        );
+        LoginController controller = new LoginController(stormpathInternalConfig(), loginErrorModelFactory);
 
         return createSpaAwareSpringController(controller);
     }
@@ -859,11 +931,7 @@ public abstract class AbstractStormpathWebMvcConfiguration {
             return createIdSiteController(idSiteForgotUri);
         }
 
-        ForgotPasswordController controller = new ForgotPasswordController(
-                stormpathForgotPasswordControllerConfigResolver(),
-                stormpathLoginControllerConfigResolver().getUri(),
-                stormpathAccountStoreResolver()
-        );
+        ForgotPasswordController controller = new ForgotPasswordController(stormpathInternalConfig());
 
         return createSpaAwareSpringController(controller);
     }
@@ -1042,14 +1110,10 @@ public abstract class AbstractStormpathWebMvcConfiguration {
 
         //otherwise standard registration:
         RegisterController controller = new RegisterController(
-                stormpathRegisterControllerConfigResolver(),
+                stormpathInternalConfig(),
                 client,
                 stormpathRequestEventPublisher(),
-                toDefaultFields(stormpathRegisterFormFields()),
-                stormpathAuthenticationResultSaver(),
-                stormpathLoginControllerConfigResolver().getUri(),
-                stormpathVerifyControllerConfigResolver().getView()
-
+                toDefaultFields(stormpathRegisterFormFields())
         );
 
         return createSpaAwareSpringController(controller);
@@ -1085,11 +1149,7 @@ public abstract class AbstractStormpathWebMvcConfiguration {
             return createIdSiteController(null);
         }
 
-        SendVerificationEmailController controller = new SendVerificationEmailController(
-                stormpathSendVerificationEmailControllerConfigResolver(),
-                stormpathLoginControllerConfigResolver().getUri(),
-                stormpathAccountStoreResolver()
-        );
+        SendVerificationEmailController controller = new SendVerificationEmailController(stormpathInternalConfig());
 
         return createSpringController(controller);
     }
@@ -1104,11 +1164,7 @@ public abstract class AbstractStormpathWebMvcConfiguration {
             return createIdSiteController(null);
         }
 
-        ChangePasswordController controller = new ChangePasswordController(
-                stormpathChangePasswordControllerConfigResolver(),
-                stormpathForgotPasswordControllerConfigResolver().getUri(),
-                stormpathLoginControllerConfigResolver().getUri()
-        );
+        ChangePasswordController controller = new ChangePasswordController(stormpathInternalConfig());
 
         return createSpaAwareSpringController(controller);
     }
@@ -1224,6 +1280,189 @@ public abstract class AbstractStormpathWebMvcConfiguration {
 
     public Controller stormpathLogoutController() {
         return createSpringController(stormpathMvcLogoutController());
+    }
+
+    public Config stormpathInternalConfig() {
+
+        final com.stormpath.sdk.servlet.i18n.MessageSource messageSource = stormpathMessageSource();
+        final Resolver<Locale> localeResolver = stormpathLocaleResolver();
+
+        return new Config() {
+            @Override
+            public ControllerConfigResolver getLoginControllerConfig() {
+                return stormpathLoginControllerConfigResolver();
+            }
+
+            @Override
+            public ControllerConfigResolver getLogoutControllerConfig() {
+                return stormpathLogoutControllerConfigResolver();
+            }
+
+            @Override
+            public ControllerConfigResolver getRegisterControllerConfig() {
+                return stormpathRegisterControllerConfigResolver();
+            }
+
+            @Override
+            public ControllerConfigResolver getForgotPasswordControllerConfig() {
+                return stormpathForgotPasswordControllerConfigResolver();
+            }
+
+            @Override
+            public ControllerConfigResolver getVerifyControllerConfig() {
+                return stormpathVerifyControllerConfigResolver();
+            }
+
+            @Override
+            public ControllerConfigResolver getSendVerificationEmailControllerConfig() {
+                return stormpathSendVerificationEmailControllerConfigResolver();
+            }
+
+            @Override
+            public ControllerConfigResolver getChangePasswordControllerConfig() {
+                return stormpathChangePasswordControllerConfigResolver();
+            }
+
+            @Override
+            public Saver<AuthenticationResult> getAuthenticationResultSaver() {
+                return stormpathAuthenticationResultSaver();
+            }
+
+            @Override
+            public AccountStoreResolver getAccountStoreResolver() {
+                return stormpathAccountStoreResolver();
+            }
+
+            @Override
+            public boolean isLogoutInvalidateHttpSession() {
+                return logoutInvalidateHttpSession;
+            }
+
+            @Override
+            public String getAccessTokenUrl() {
+                return accessTokenUri;
+            }
+
+            @Override
+            public String getUnauthorizedUrl() {
+                return "/unauthorized";
+            }
+
+            @Override
+            public boolean isMeEnabled() {
+                return meEnabled;
+            }
+
+            @Override
+            public String getMeUrl() {
+                return meUri;
+            }
+
+            @Override
+            public boolean getMeExpandGroups() {
+                return meExpandGroups;
+            }
+
+            @Override
+            public CookieConfig getAccountCookieConfig() {
+                return stormpathAccountCookieConfig();
+            }
+
+            @Override
+            public long getAccountJwtTtl() {
+                return accountJwtTtl;
+            }
+
+            @Override
+            public String getAccessTokenValidationStrategy() {
+                return accessTokenValidationStrategy;
+            }
+
+            @Override
+            public <T> T getInstance(String classPropertyName) throws ServletException {
+                if (MessageTag.LOCALE_RESOLVER_CONFIG_KEY.equals(classPropertyName)) {
+                    return (T) localeResolver;
+                } else if (MessageTag.MESSAGE_SOURCE_CONFIG_KEY.equals(classPropertyName)) {
+                    return (T) messageSource;
+                } else {
+                    String msg = "The config key '" + classPropertyName + "' is not supported in Spring environments " +
+                            "- inject the required dependency via Spring config (e.g. @Autowired) instead.";
+                    throw new UnsupportedOperationException(msg);
+                }
+            }
+
+            @Override
+            public <T> Map<String, T> getInstances(String propertyNamePrefix, Class<T> expectedType)
+                    throws ServletException {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public List<String> getProducedMediaTypes() {
+                return produces;
+            }
+
+            @Override
+            public int size() {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public boolean isEmpty() {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public boolean containsKey(Object o) {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public boolean containsValue(Object o) {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public String get(Object o) {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public String put(String s, String s2) {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public String remove(Object o) {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public void putAll(Map<? extends String, ? extends String> map) {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public void clear() {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public Set<String> keySet() {
+                //The Spring Boot WebMVC + Spring Security Example causes this method to be invoked. Thus, we cannot throw an exception here.
+                return java.util.Collections.EMPTY_SET;
+            }
+
+            @Override
+            public Collection<String> values() {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+
+            @Override
+            public Set<Entry<String, String>> entrySet() {
+                throw new UnsupportedOperationException("Not supported for spring environments.");
+            }
+        };
     }
 
     public FilterChainResolver stormpathFilterChainResolver() {
